prompt-builder-local/mac-installer/requirements_local_only.txt
PyQt6>=6.4.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.20.0
Pillow>=9.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
psutil>=5.9.0
llama-cpp-python>=0.2.0